requirements.txt
python-dotenv
python-teamwork